# resources/swagger/openapi.yaml
openapi: 3.1.0
info:
  title: BookSwap
  version: 0.0.1
servers:
  - url: http://localhost/api/v1
security:
  - http: [ ]
paths:
...
  /authors/{author}:
    parameters:
      - $ref: "components/parameters/author.yaml"
    get:
      operationId: authors.show
      tags:
        - Author
      security: [ ]
      responses:
        "200":
          $ref: "components/responses/api/Author_Response.yaml"
        "400":
          $ref: "components/responses/exceptions/BadRequest.yaml"
        "404":
          $ref: "components/responses/exceptions/ModelNotFoundException.yaml"
    patch:
      operationId: authors.update
      tags:
        - Author
      requestBody:
        $ref: "components/requestBodies/UpdateAuthorRequest_Body.yaml"
      responses:
        "200":
          $ref: "components/responses/api/Author_Response.yaml"
        "400":
          $ref: "components/responses/exceptions/BadRequest.yaml"
        "401":
          $ref: "components/responses/exceptions/AuthenticationException.yaml"
        "403":
          $ref: "components/responses/exceptions/Forbidden.yaml"
        "422":
          $ref: "components/responses/exceptions/ValidationException.yaml"
    delete:
      operationId: authors.destroy
      tags:
        - Author
      responses:
        "202":
          $ref: "components/responses/api/Accepted_Response.yaml"
        "400":
          $ref: "components/responses/exceptions/BadRequest.yaml"
        "401":
          $ref: "components/responses/exceptions/AuthenticationException.yaml"
        "403":
          $ref: "components/responses/exceptions/Forbidden.yaml"
...
  /books/{book}:
    parameters:
      - $ref: "components/parameters/book.yaml"
    get:
      operationId: books.show
      tags:
        - Book
      security: [ ]
      responses:
        "200":
          $ref: "components/responses/api/Book_Response.yaml"
        "400":
          $ref: "components/responses/exceptions/BadRequest.yaml"
        "404":
          $ref: "components/responses/exceptions/ModelNotFoundException.yaml"
    patch:
      operationId: books.update
      tags:
        - Book
      requestBody:
        $ref: "components/requestBodies/UpdateBookRequest_Body.yaml"
      responses:
        "200":
          $ref: "components/responses/api/Book_Response.yaml"
        "400":
          $ref: "components/responses/exceptions/BadRequest.yaml"
        "401":
          $ref: "components/responses/exceptions/AuthenticationException.yaml"
        "403":
          $ref: "components/responses/exceptions/Forbidden.yaml"
        "404":
          $ref: "components/responses/exceptions/ModelNotFoundException.yaml"
        "422":
          $ref: "components/responses/exceptions/ValidationException.yaml"
    delete:
      operationId: books.destroy
      tags:
        - Book
      responses:
        "202":
          $ref: "components/responses/api/Accepted_Response.yaml"
        "400":
          $ref: "components/responses/exceptions/BadRequest.yaml"
        "401":
          $ref: "components/responses/exceptions/AuthenticationException.yaml"
        "403":
          $ref: "components/responses/exceptions/Forbidden.yaml"
...
  /book_types/{book_type}:
    parameters:
      - $ref: "components/parameters/book_type.yaml"
    get:
      operationId: book_types.show
      tags:
        - BookType
      security: [ ]
      responses:
        "200":
          $ref: "components/responses/api/BookType_Response.yaml"
        "400":
          $ref: "components/responses/exceptions/BadRequest.yaml"
        "404":
          $ref: "components/responses/exceptions/ModelNotFoundException.yaml"
    patch:
      operationId: book_types.update
      tags:
        - BookType
      requestBody:
        $ref: "components/requestBodies/UpdateBookTypeRequest_Body.yaml"
      responses:
        "200":
          $ref: "components/responses/api/BookType_Response.yaml"
        "400":
          $ref: "components/responses/exceptions/BadRequest.yaml"
        "401":
          $ref: "components/responses/exceptions/AuthenticationException.yaml"
        "403":
          $ref: "components/responses/exceptions/Forbidden.yaml"
        "404":
          $ref: "components/responses/exceptions/ModelNotFoundException.yaml"
        "422":
          $ref: "components/responses/exceptions/ValidationException.yaml"
    delete:
      operationId: book_types.destroy
      tags:
        - BookType
      responses:
        "202":
          $ref: "components/responses/api/Accepted_Response.yaml"
        "400":
          $ref: "components/responses/exceptions/BadRequest.yaml"
        "401":
          $ref: "components/responses/exceptions/AuthenticationException.yaml"
        "403":
          $ref: "components/responses/exceptions/Forbidden.yaml"
...
  /chats/{chat}:
    parameters:
      - $ref: "components/parameters/chat.yaml"
    get:
      operationId: chats.show
      tags:
        - Chat
      responses:
        "200":
          $ref: "components/responses/api/Chat_Response.yaml"
        "400":
          $ref: "components/responses/exceptions/BadRequest.yaml"
        "401":
          $ref: "components/responses/exceptions/AuthenticationException.yaml"
        "403":
          $ref: "components/responses/exceptions/Forbidden.yaml"
        "404":
          $ref: "components/responses/exceptions/ModelNotFoundException.yaml"
    delete:
      operationId: chats.destroy
      tags:
        - Chat
      responses:
        "202":
          $ref: "components/responses/api/Accepted_Response.yaml"
        "400":
          $ref: "components/responses/exceptions/BadRequest.yaml"
        "401":
          $ref: "components/responses/exceptions/AuthenticationException.yaml"
        "403":
          $ref: "components/responses/exceptions/Forbidden.yaml"
...
  /covers/{cover}:
    parameters:
      - $ref: "components/parameters/cover.yaml"
    get:
      operationId: covers.show
      tags:
        - Cover
      security: [ ]
      responses:
        "200":
          $ref: "components/responses/api/Cover_Response.yaml"
        "400":
          $ref: "components/responses/exceptions/BadRequest.yaml"
        "404":
          $ref: "components/responses/exceptions/ModelNotFoundException.yaml"
    delete:
      operationId: covers.destroy
      tags:
        - Cover
      responses:
        "202":
          $ref: "components/responses/api/Accepted_Response.yaml"
        "400":
          $ref: "components/responses/exceptions/BadRequest.yaml"
        "401":
          $ref: "components/responses/exceptions/AuthenticationException.yaml"
        "403":
          $ref: "components/responses/exceptions/Forbidden.yaml"
...
  /trade_offers/{trade_offer}:
    parameters:
      - $ref: "components/parameters/trade_offer.yaml"
    get:
      operationId: trade_offers.show
      tags:
        - TradeOffer
      responses:
        "200":
          $ref: "components/responses/api/TradeOffer_Response.yaml"
        "400":
          $ref: "components/responses/exceptions/BadRequest.yaml"
        "401":
          $ref: "components/responses/exceptions/AuthenticationException.yaml"
        "403":
          $ref: "components/responses/exceptions/Forbidden.yaml"
        "404":
          $ref: "components/responses/exceptions/ModelNotFoundException.yaml"
    patch:
      operationId: trade_offers.update
      tags:
        - TradeOffer
      requestBody:
        $ref: "components/requestBodies/UpdateTradeOfferRequest_Body.yaml"
      responses:
        "200":
          $ref: "components/responses/api/TradeOffer_Response.yaml"
        "400":
          $ref: "components/responses/exceptions/BadRequest.yaml"
        "401":
          $ref: "components/responses/exceptions/AuthenticationException.yaml"
        "403":
          $ref: "components/responses/exceptions/Forbidden.yaml"
        "404":
          $ref: "components/responses/exceptions/ModelNotFoundException.yaml"
        "422":
          $ref: "components/responses/exceptions/ValidationException.yaml"
    delete:
      operationId: trade_offers.destroy
      tags:
        - TradeOffer
      responses:
        "202":
          $ref: "components/responses/api/Accepted_Response.yaml"
        "400":
          $ref: "components/responses/exceptions/BadRequest.yaml"
        "401":
          $ref: "components/responses/exceptions/AuthenticationException.yaml"
        "403":
          $ref: "components/responses/exceptions/Forbidden.yaml"
        "404":
          $ref: "components/responses/exceptions/ModelNotFoundException.yaml"
...
  /filters/{filter}:
    parameters:
      - $ref: "components/parameters/filter.yaml"
    get:
      operationId: filters.show
      tags:
        - Filter
      security: [ ]
      responses:
        "200":
          $ref: "components/responses/api/Filter_Response.yaml"
        "400":
          $ref: "components/responses/exceptions/BadRequest.yaml"
        "404":
          $ref: "components/responses/exceptions/ModelNotFoundException.yaml"
    patch:
      operationId: filters.update
      tags:
        - Filter
      requestBody:
        $ref: "components/requestBodies/UpdateFilterRequest_Body.yaml"
      responses:
        "200":
          $ref: "components/responses/api/Filter_Response.yaml"
        "400":
          $ref: "components/responses/exceptions/BadRequest.yaml"
        "401":
          $ref: "components/responses/exceptions/AuthenticationException.yaml"
        "403":
          $ref: "components/responses/exceptions/Forbidden.yaml"
        "404":
          $ref: "components/responses/exceptions/ModelNotFoundException.yaml"
        "422":
          $ref: "components/responses/exceptions/ValidationException.yaml"
    delete:
      operationId: filters.destroy
      tags:
        - Filter
      responses:
        "202":
          $ref: "components/responses/api/Accepted_Response.yaml"
        "400":
          $ref: "components/responses/exceptions/BadRequest.yaml"
        "401":
          $ref: "components/responses/exceptions/AuthenticationException.yaml"
        "403":
          $ref: "components/responses/exceptions/Forbidden.yaml"
...
  /genres/{genre}:
    parameters:
      - $ref: "components/parameters/genre.yaml"
    get:
      operationId: genres.show
      tags:
        - Genre
      security: [ ]
      responses:
        "200":
          $ref: "components/responses/api/Genre_Response.yaml"
        "400":
          $ref: "components/responses/exceptions/BadRequest.yaml"
        "404":
          $ref: "components/responses/exceptions/ModelNotFoundException.yaml"
    patch:
      operationId: genres.update
      tags:
        - Genre
      requestBody:
        $ref: "components/requestBodies/UpdateGenreRequest_Body.yaml"
      responses:
        "200":
          $ref: "components/responses/api/Genre_Response.yaml"
        "400":
          $ref: "components/responses/exceptions/BadRequest.yaml"
        "401":
          $ref: "components/responses/exceptions/AuthenticationException.yaml"
        "403":
          $ref: "components/responses/exceptions/Forbidden.yaml"
        "404":
          $ref: "components/responses/exceptions/ModelNotFoundException.yaml"
        "422":
          $ref: "components/responses/exceptions/ValidationException.yaml"
    delete:
      operationId: genres.destroy
      tags:
        - Genre
      responses:
        "202":
          $ref: "components/responses/api/Accepted_Response.yaml"
        "400":
          $ref: "components/responses/exceptions/BadRequest.yaml"
        "401":
          $ref: "components/responses/exceptions/AuthenticationException.yaml"
        "403":
          $ref: "components/responses/exceptions/Forbidden.yaml"
...
  /photos/{photo}:
    parameters:
      - $ref: "components/parameters/photo.yaml"
    get:
      operationId: photos.show
      tags:
        - Photo
      security: [ ]
      responses:
        "200":
          $ref: "components/responses/api/Photo_Response.yaml"
        "400":
          $ref: "components/responses/exceptions/BadRequest.yaml"
        "404":
          $ref: "components/responses/exceptions/ModelNotFoundException.yaml"
    delete:
      operationId: photos.destroy
      tags:
        - Photo
      responses:
        "202":
          $ref: "components/responses/api/Accepted_Response.yaml"
        "400":
          $ref: "components/responses/exceptions/BadRequest.yaml"
        "401":
          $ref: "components/responses/exceptions/AuthenticationException.yaml"
        "403":
          $ref: "components/responses/exceptions/Forbidden.yaml"
...
  /settings/{setting}:
    parameters:
      - $ref: "components/parameters/setting.yaml"
    get:
      operationId: settings.show
      tags:
        - Setting
      description: "Требуется аутентификация"
      security:
        - http: [ ]
      responses:
        "200":
          $ref: "components/responses/api/Setting_Response.yaml"
        "400":
          $ref: "components/responses/exceptions/BadRequest.yaml"
        "401":
          $ref: "components/responses/exceptions/AuthenticationException.yaml"
        "403":
          $ref: "components/responses/exceptions/Forbidden.yaml"
        "404":
          $ref: "components/responses/exceptions/ModelNotFoundException.yaml"
    patch:
      operationId: settings.update
      tags:
        - Setting
      requestBody:
        $ref: "components/requestBodies/UpdateSettingRequest_Body.yaml"
      responses:
        "200":
          $ref: "components/responses/api/Setting_Response.yaml"
        "400":
          $ref: "components/responses/exceptions/BadRequest.yaml"
        "401":
          $ref: "components/responses/exceptions/AuthenticationException.yaml"
        "403":
          $ref: "components/responses/exceptions/Forbidden.yaml"
        "404":
          $ref: "components/responses/exceptions/ModelNotFoundException.yaml"
        "422":
          $ref: "components/responses/exceptions/ValidationException.yaml"
    delete:
      operationId: settings.destroy
      tags:
        - Setting
      responses:
        "202":
          $ref: "components/responses/api/Accepted_Response.yaml"
        "400":
          $ref: "components/responses/exceptions/BadRequest.yaml"
        "401":
          $ref: "components/responses/exceptions/AuthenticationException.yaml"
        "403":
          $ref: "components/responses/exceptions/Forbidden.yaml"
...
  /users/{user}:
    parameters:
      - $ref: "components/parameters/user.yaml"
    get:
      operationId: users.show
      tags:
        - User
      security: [ ]
      responses:
        "200":
          $ref: "components/responses/api/User_Response.yaml"
        "400":
          $ref: "components/responses/exceptions/BadRequest.yaml"
        "404":
          $ref: "components/responses/exceptions/ModelNotFoundException.yaml"
    patch:
      operationId: users.update
      tags:
        - User
      requestBody:
        $ref: "components/requestBodies/UpdateUserRequest_Body.yaml"
      responses:
        "200":
          $ref: "components/responses/api/User_Response.yaml"
        "400":
          $ref: "components/responses/exceptions/BadRequest.yaml"
        "401":
          $ref: "components/responses/exceptions/AuthenticationException.yaml"
        "403":
          $ref: "components/responses/exceptions/Forbidden.yaml"
        "404":
          $ref: "components/responses/exceptions/ModelNotFoundException.yaml"
        "422":
          $ref: "components/responses/exceptions/ValidationException.yaml"
    delete:
      operationId: users.destroy
      tags:
        - User
      responses:
        "202":
          $ref: "components/responses/api/Accepted_Response.yaml"
        "400":
          $ref: "components/responses/exceptions/BadRequest.yaml"
        "401":
          $ref: "components/responses/exceptions/AuthenticationException.yaml"
        "403":
          $ref: "components/responses/exceptions/Forbidden.yaml"
...
  /me/dislike/{book}:
    parameters:
      - $ref: "components/parameters/book.yaml"
    patch:
      operationId: me.favorites.dislike
      tags:
        - Me
      responses:
        "202":
          $ref: "components/responses/api/Accepted_Response.yaml"
        "400":
          $ref: "components/responses/exceptions/BadRequest.yaml"
        "401":
          $ref: "components/responses/exceptions/AuthenticationException.yaml"
        "403":
          $ref: "components/responses/exceptions/Forbidden.yaml"
        "404":
          $ref: "components/responses/exceptions/ModelNotFoundException.yaml"
...
  /me/trade_history:
    get:
      operationId: me.tradeHistory
      tags:
        - Me
      responses:
        "200":
          $ref: "components/responses/api/TradeOfferHistory_Response.yaml"
        "400":
          $ref: "components/responses/exceptions/BadRequest.yaml"
        "401":
          $ref: "components/responses/exceptions/AuthenticationException.yaml"
        "403":
          $ref: "components/responses/exceptions/Forbidden.yaml"
...
components:
  securitySchemes:
    http:
      type: http
      scheme: bearer